requirements.txt
numpy>=1.26.3
Pillow>=10.2.0
torch>=2.1.2
transformers>=4.42.0
scipy>=1.12.0
soundfile>=0.12.1
pandas>=2.1.0
sentencepiece>=0.1.99
accelerate>=0.25.0
timm
opencv-python
